Add check to see if user has permissions to modify or create

+++ pages/api/delete/[id].ts
@@ -1,6 +1,8 @@
 import { getSession } from "next-auth/react";
 import type { NextApiRequest, NextApiResponse } from "next";
 import prisma from "../../../lib/prisma";
+import { getUser } from "../../../lib/getUser";
+import { revalidateNow } from "../../../lib/revalidateHelper";
 
 // POST /api/post
 // Required fields in body: title
@@ -12,32 +14,35 @@ export default async function handle(
   const session = await getSession({ req });
   const postId = parseInt(req.query.id as string);
 
+  let userTemp;
   if (session) {
-    // Author has bug, but why?
-    // const author = { connect: { email: session?.user?.email } };
     try {
-      console.log("postId")
-      console.log(postId)
-      console.log("Trying...");
-      const result = await prisma.post.delete({
-        where: { id: postId },
-      });
-      console.log("result");
-      console.log(result);
-      res.status(200).json(result);
+      let user = await getUser(session?.user?.email);
+      userTemp = user;
+      if (!user[0].authorizedToPublish) {
+        throw 402;
+      } else {
+        // Author has bug, but why?
+        // const author = { connect: { email: session?.user?.email } };
+        const result = await prisma.post.delete({
+          where: { id: postId },
+        });
+        res.status(200).json(result);
+      }
     } catch (error) {
-      res.status(500).send({
-        message: "Database Error",
-      });
+      if (error === 402) {
+        res.status(200).json({
+          error: `You do not have the permission to delete recipes.`,
+        });
+        res.status(500).send({
+          message: "Database Error",
+        });
+      }
     }
   }
-
   if (!session) {
     res.status(403).send({
       message: "You need to be logged in to perform this action.",
     });
   }
-
-  // console.log("result");
-  // console.log(session);
 }
